fix(PlainTemplate): avoid to duplicate information for Object properties
object properties doesn't need to show the label because the PlainTemplate already take in
consideration the title within the children to be rendered, same happens for the description. In
that sense, if we render the label and description for objects we will end up by re rendering
structures and information twice.