feat(document-generation): add base to develop service
use base of "puppeteer-lambda-starter-kit" to develop a custom solution for the structure of this
service using "chrome-aws-lambda" and "html-pdf-chrome". This base will simple use puppeteer to
return the title of the website depending on a given url, besides, it includes the templates used by
the old platform in order to generate documents.

re #36